Move CQD Score tests
